updated the first function
I think this version actually spits out the true distance...(I had a conceptual flaw in my thinking).
I tried to summarize everything in a one line list comprehension (below), maybe you could judge if I did it correctly? it seems odd, I need to assign the result of the function to a new variable in order to make it work ( I didn't need to do that with the first solution).